% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignFolds.R
\name{AssignFolds}
\alias{AssignFolds}
\title{Assigning CV Folds}
\usage{
AssignFolds(
  data,
  spatial = TRUE,
  k = 10,
  species = NULL,
  coords = c("lon", "lat"),
  crs = 4326,
  fold_col = "Folds"
)
}
\arguments{
\item{data}{data frame containing species observation data and spatial coordinates}

\item{spatial}{logical; if TRUE (default), uses blockCV to create spatial block folds. If FALSE, assigns random folds}

\item{k}{integer; number of cross-validation folds to create (default is 10)}

\item{species}{character; optional column name for species presence/absence or abundance data to balance occurrences across spatial folds}

\item{coords}{character vector of length 2; column names representing x/longitude and y/latitude coordinates (default is c("lon", "lat"))}

\item{crs}{integer or character; Coordinate Reference System (CRS) for spatial coordinates (default is 4326 / WGS84)}

\item{fold_col}{character; column name in the returned data frame where assigned fold identifiers will be stored (default is "Folds")}
}
\value{
data frame corresponding to input data with an additional column containing assigned fold letters
}
\description{
Assign spatial or non-spatial cross-validation folds to a dataset for species distribution modeling.
}
